Create sonicwall_commands.txt
An example of the sonicwall_commands.txt file that the Invoke-RunSonicwallCommandScript.txt file would take. Replace any of the items in <> with their respective values. This setup would create address and service groups, then create a nat-rule and access-rule for allowing a port-forward on a sonicwall from WAN to LAN.